split up large files in MongoDBTests
